+++ coreservices/partsrelationshipservice/integration-tests/src/test/java/net/catenax/prs/integrationtest/GetPartsTreeByObjectIdIntegrationTests.java
@@ -7,7 +7,7 @@
 // See the LICENSE file(s) distributed with this work for
 // additional information regarding license terms.
 //
-package net.catenax.prs.test;
+package net.catenax.prs.integrationtest;
 
 import com.catenax.partsrelationshipservice.dtos.PartRelationshipsWithInfos;
 import com.catenax.partsrelationshipservice.dtos.PartsTreeView;
